Собрал сборку, работаю с рефлексией
