+++ FileManager/FileManager/Core/Core.cs
@@ -1,147 +1,197 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Reflection;
 using System.Text;
 
 namespace FileManager
 {
     public static class Core
     {
-
+        private static List<IDirective> _directives= new List<IDirective>();
 
 
 
         #region For delete
 
-        ///// <summary>
-        ///// Основной метод, который запускает бесконечный цикл работы программы
-        ///// </summary>
-        public static void RunConsole()
-        {
-            bool exit = false;
+        /////// <summary>
+        /////// Основной метод, который запускает бесконечный цикл работы программы
+        /////// </summary>
+        //public static void RunConsole()
+        //{
+        //    bool exit = false;
 
-            UserCommands.CommandInt = -1;
+        //    UserCommands.CommandInt = -1;
 
-            Engine.StartCommandExecuter();
+        //    StartCommandExecuter();
 
-            while (!exit)
-            {
-                PseudoConsoleUI.SetCursorPosition(UserCommands.DrivesAndDirs);
+        //    while (!exit)
+        //    {
+        //        PseudoConsoleUI.SetCursorPosition(UserCommands.DrivesAndDirs);
 
-                Engine.UserCommandReader();
+        //        UserCommandReader();
 
-                string[] commandsStringArray = UserCommands.Command.Split();
+        //        string[] commandsStringArray = UserCommands.Command.Split();
 
-                if (UserCommands.CommandInt == -1)
-                {
-                    continue;
-                }
-                switch (Engine.CommandInt)
-                {
-                    case 0:
-                        {
-                            exit = Engine.ExitCommandExecuter();
+        //        if (UserCommands.CommandInt == -1)
+        //        {
+        //            continue;
+        //        }
+        //        switch (Engine.CommandInt)
+        //        {
+        //            case 0:
+        //                {
+        //                    exit = Engine.ExitCommandExecuter();
 
-                            break;
-                        }
-                    //case 1:
-                    //    {
-                    //    Engine.ChangeDirectoryCommandExecuter(commandsStringArray[1]);
+        //                    break;
+        //                }
+        //case 1:
+        //    {
+        //    Engine.ChangeDirectoryCommandExecuter(commandsStringArray[1]);
 
-                    //    PseudoConsoleUI.PrintDirectoryProrerties(Engine.drivesAndDirectories);
+        //    PseudoConsoleUI.PrintDirectoryProrerties(Engine.drivesAndDirectories);
 
-                    //    break;
-                    //}
-                    //case 2:
-                    //    {
-                    //    Engine.ClearConsoleCommandExecuter();
+        //    break;
+        //}
+        //case 2:
+        //    {
+        //    Engine.ClearConsoleCommandExecuter();
 
-                    //    break;
-                    //}
-                    //case 3:
-                    //    {
-                    //    PseudoConsoleUI.ShowCopyCommand(Engine.CopyCommandExecuter());
+        //    break;
+        //}
+        //case 3:
+        //    {
+        //    PseudoConsoleUI.ShowCopyCommand(Engine.CopyCommandExecuter());
 
-                    //    break;
-                    //}
-                    //case 4:
-                    //    {
-                    //    Engine.DeleteCommandExecuter();
+        //    break;
+        //}
+        //case 4:
+        //    {
+        //    Engine.DeleteCommandExecuter();
 
-                    //    break;
-                    //}
-                    //case 5:
-                    //    {
-                    //    Engine.DeleteTreeCommandExecuter();
+        //    break;
+        //}
+        //case 5:
+        //    {
+        //    Engine.DeleteTreeCommandExecuter();
 
-                    //    break;
-                    //}
-                    //case 6:
-                    //    {
-                    //    Engine.MoveCommandExecuter();
+        //    break;
+        //}
+        //case 6:
+        //    {
+        //    Engine.MoveCommandExecuter();
 
-                    //    break;
-                    //}
-                    //case 7:
-                    //    {
-                    //    Engine.MakingDirectoryCommandExecuter();
+        //    break;
+        //}
+        //case 7:
+        //    {
+        //    Engine.MakingDirectoryCommandExecuter();
 
-                    //    break;
-                    //}
-                    //case 8:
-                    //    {
-                    //    Engine.RemoveDirectoryCommandExecuter();
+        //    break;
+        //}
+        //case 8:
+        //    {
+        //    Engine.RemoveDirectoryCommandExecuter();
 
-                    //    break;
-                    //}
-                    //case 9:
-                    //    {
-                    //    Engine.ShowAllSubdirectoriesAndFilesByPages();
+        //    break;
+        //}
+        //case 9:
+        //    {
+        //    Engine.ShowAllSubdirectoriesAndFilesByPages();
 
-                    //    PseudoConsoleUI.PrintDirectoryProrerties(Engine.drivesAndDirectories);
+        //    PseudoConsoleUI.PrintDirectoryProrerties(Engine.drivesAndDirectories);
 
-                    //    break;
-                    //}
-                    //case 10:
-                    //    {
-                    //    PseudoConsoleUI.WriteHelp(Engine.HelpCommandExecuter());
+        //    break;
+        //}
+        //case 10:
+        //    {
+        //    PseudoConsoleUI.WriteHelp(Engine.HelpCommandExecuter());
 
-                    //    break;
-                    //}
-                    //case 11:
-                    //    {
-                    //    Engine.FilePropertiesCommandExecuter();
+        //    break;
+        //}
+        //case 11:
+        //    {
+        //    Engine.FilePropertiesCommandExecuter();
 
-                    //    break;
-                    //}
-                    //case 12:
-                    //    {
-                    //    Engine.ProcessCommandExecuter();
+        //    break;
+        //}
+        //case 12:
+        //    {
+        //    Engine.ProcessCommandExecuter();
 
-                    //    break;
-                    //}
-                    default:
-                        {
-                            break;
-                        }
+        //    break;
+        //}
+        //            default:
+        //                {
+        //                    break;
+        //                }
 
-                }
-                UserCommands.CommandInt = -1;
-            }
-        }
+        //        }
+        //        UserCommands.CommandInt = -1;
+        //    }
+        //}
 
         #endregion
 
-        //    public static void RunConsole()
-        //{
-        //}
+        public static void RunConsole()
+        {
+            DirectiveSelection();
+        }
+
 
 
 
 
         /// <summary>
+        /// Метод запускает цикл по выбору и выполнению задачи
+        /// </summary>
+        private static void DirectiveSelection()
+        {
+            bool exit = false;
+
+            StartCommandExecuter();
+
+            Directives();
+
+            while (exit)
+            {
+                PseudoConsoleUI.SetCursorPosition(UserCommands.DrivesAndDirs);
+
+                UserCommands.Command = Console.ReadLine();
+
+                if (UserCommands.Command.Contains(".EXE") || UserCommands.Command.Contains(".COM") || UserCommands.Command.Contains(".BAT"))
+                {
+                    UserCommands.Command = String.Format(UserCommands.DrivesAndDirs.CuttentDirectory.ToString() + UserCommands.Command);
+
+                    BasicLogic.CreateProcess(UserCommands.Command);
+
+                    continue;
+                }
+                string[] commandsStringArray = UserCommands.Command.Split();
+
+                //if (UserCommands.CommandInt == -1)
+                //{
+                //    continue;
+                //}
+                
+                //userChoice = Console.ReadLine();
+
+                foreach (IDirective directive in _directives)
+                {
+                    if (directive.DirectiveName == commandsStringArray[0].ToUpperInvariant())
+                    {
+                        directive.RunDirective();
+                    }
+                }
+            }
+        }
+
+        
+        /// <summary>
         /// Метод при помощи механизмов класса System.Reflection динамически подключает библиотеку классов
         /// </summary>
-        private static void Tasks()
+        private static void Directives()
         {
             Assembly asm = Assembly.GetExecutingAssembly();
 
@@ -151,7 +201,7 @@ namespace FileManager
             {
                 FieldInfo field = type.GetField("_directiveName");
 
-                if (field != null)
+                if (type.IsInterface && type.GetField("_directiveName")!=null)
                 {
                     object obj = Activator.CreateInstance(type);
 
@@ -160,54 +210,63 @@ namespace FileManager
             }
         }
 
+
+
         /// <summary>
-        /// Метод запускает цикл по выбору и выполнению задачи
+        /// Метод запускает консоль, читает из файла сохраненный при прошлом корректном выходе статус.
         /// </summary>
-        private static void TaskSelection()
+        public static void StartCommandExecuter()
         {
-            string userChoice = "";
-
-            Tasks();
-
-            while (userChoice != "0")
+            if (File.Exists("path.json"))
             {
-                Console.Clear();
+                string temp = File.ReadAllText("path.json");
 
-                Console.WriteLine($"Список практических работ по дисциплине \"Алгоритмы и структуры данных\"");
+                string[] tempArr = JsonConvert.DeserializeObject<string>(temp).Split("|W|");
 
-                foreach (IDirective directive in _directives)
+                UserCommands.DrivesAndDirs.CurrentDrive = new DriveInfo(tempArr[0]);
+
+                UserCommands.DrivesAndDirs.CuttentDirectory = new DirectoryInfo(tempArr[1]);
+
+                if (!UserCommands.DrivesAndDirs.CuttentDirectory.Exists)
                 {
-                    Console.WriteLine(directive.DirectiveName);
+                    UserCommands.DrivesAndDirs = new DrivesAndDirectories();
                 }
-                Console.WriteLine($"Введите номер задачи или 0 для выхода");
-
-                userChoice = Console.ReadLine();
-
-                foreach (IDirective directive in _directives)
-                {
-                    if (directive.DirectiveName == userChoice.ToUpperInvariant())
-                    {
-                        directive.RunDirective();
-
-                        PressToExit();
-                    }
-                }
+            }
+            else
+            {
+                UserCommands.DrivesAndDirs = new DrivesAndDirectories();
             }
         }
 
+
+
+
         /// <summary>
-        /// Метод запрашивает нажатия любой кнопки
+        /// Метод считывает из консоли команды пользователя и сравнивает их со списком команд
         /// </summary>
-        private static void PressToExit()
+        public static void UserCommandReader()
         {
-            Console.WriteLine($"Press any key to exit...");
+            UserCommands.Command = Console.ReadLine();
 
-            Console.ReadKey();
+            //string comm = UserCommands.Command != "" ? UserCommands.Command.Split()[0] : "";
+
+            //for (int i = 0; i < commands.Count; i++)
+            //{
+            //    //if (commands[i] == comm)
+                //{
+                //    UserCommands.CommandInt = i;
+
+                //    break;
+                //}
+                //else 
+                //if (comm.Contains(".EXE") || comm.Contains(".COM") || comm.Contains(".BAT"))
+                //{
+                //    comm = String.Format(UserCommands.DrivesAndDirs.CuttentDirectory.ToString() + UserCommands.Command);
+
+                //    BasicLogic.CreateProcess(comm);
+                //}
+            //}
         }
-
-
-
-
