+++ FileManager/FileManager/Logic/DirectiveHELP.cs
@@ -13,7 +13,7 @@ namespace FileManager
 
         public void RunDirective(params string[] args)
         {
-            StartDirective(args[0]);
+            //StartDirective(args[0]);
         }
 
         /// <summary>
